[Optimize] Optimize doc and CI bash, Refactoring OALRuntime
