+++ .github/workflows/native-image.yaml
@@ -32,7 +32,7 @@ jobs:
           github-token: ${{ secrets.GITHUB_TOKEN }}
       - name: Native Compile
         run: |
-          ./mvnw -Pnative,backend clean package -Dmaven.test.skip
+          ./mvnw -Pbackend,native clean package -Dmaven.test.skip
       - uses: actions/upload-artifact@v3
         name: Upload distribution tar
         with:
